poppler::page::text_list() returns a vector of the objects, not the pointers.
Unlike immutable vectors of the embedded_file, toc_item,
the result of text_list() could be configurable in future
(resolutin, layout, annotation etc).  To call the destuctors
of the member objects during the destruction of the vector,
the vector by text_list() is changed to the vector of the
objects, instead of the pointers to the objects.